+++ config/app.config.php
@@ -3,23 +3,23 @@
 $loader = require_once(__DIR__ . "/../vendor/autoload.php");
 
 // - config slim
-$app  = new \Slim\App([
+$app = new \Slim\App([
     'settings' => [
         'displayErrorDetails' => true,
         'debug' => true
     ],
-]);
+        ]);
 
 
 // 1 - Get Slim Container Instance
-$appContainer = $app->getContainer();
+$slimContainer = $app->getContainer();
 
 // 2 - Register "Controller" in Slim Container
-$appContainer['Controller'] = function($appContainer) {
+$appContainer['Controller'] = function($slimContainer) {
 
+    $config = [];
     $service = [];
-    $config  = [];
-    $controller = new \App\Controller($config , $service);
+    $controller = new \App\Controller($config, $service);
     return $controller;
 };