+++ public/index.php
@@ -2,10 +2,11 @@
 
 $app = include(__DIR__ . '/../config/app.config.php');
 
-$app->get('/' , function($Request , $Response , $Args){
-    
-    $result = $this->Controller->action($Request , $Response , $Args);
+$app->get('/', function($Request, $Response, $Args) {
+
+    $result = $this->Controller->action();
     $Response->getBody()->write($result);
+    
 });
 
 $app->run();