+++ tests/fixture/DummyEntity.php
@@ -1,14 +1,14 @@
 <?php
-namespace Prose\Test;
+
+namespace Core\Test;
 
 /**
  * @Entity 
  * @Table(name="test_table")
- **/
+ * */
+class DummyEntity {
 
-class DummyEntity{
-    
-    /** 
+    /**
      * @Id 
      * @Column(type="integer")
      * @GeneratedValue 
@@ -16,26 +16,12 @@ class DummyEntity{
      * */
     private $id = "";
 
-    
-    /**
-     * @Column(type="string") 
-     */
-    private $label = "";
-    
     public function setId($id) {
         $this->id = $id;
     }
-    
+
     public function getId() {
         return $this->id;
     }
 
-    public function getLabel() {
-        return $this->label;
-    }
-
-    public function setLabel($label) {
-        $this->label = $label;
-        return $this;
-    }
 }