+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ### Changelog
 All notable changes to this project since v1.0.0 will be documented in this file.
 
+### v1.1.1
+- Update dependencies:
+    - **slim/slim 3.\***
+    - **phpunit-9**
+- Add namespace **Core**
+- Add Abstract class **Core::BaseController**, that acts as a basic dependencies container.
+- Add Test for **Core::BaseController** class.
+
 ### v1.1.0
 - removing Psr\Container as controller dependency.
 - renaming package to **prose/slim-3-starter**